feat: hook useRoom
